refactor: update request handling and parameter destructuring in article route handlers
