+++ src/lib/authOptions.ts
@@ -1,0 +1,36 @@
+import { AuthOptions } from "next-auth";
+import CredentialsProvider from "next-auth/providers/credentials";
+
+export const authOptions: AuthOptions = {
+  providers: [
+    CredentialsProvider({
+      name: "Credentials",
+      credentials: {
+        email: { label: "Email", type: "email" },
+        password: { label: "Password", type: "password" }
+      },
+      async authorize(credentials) {
+        const adminEmail = process.env.ADMIN_EMAIL;
+        const adminPassword = process.env.ADMIN_PASSWORD;
+
+        if (!adminEmail || !adminPassword) {
+          console.error("Admin credentials are not set.");
+          return null;
+        }
+
+        if (credentials?.email === adminEmail && credentials?.password === adminPassword) {
+          return { id: "1", name: "Admin User", email: adminEmail };
+        } else {
+          return null;
+        }
+      }
+    })
+  ],
+  pages: {
+    signIn: '/login',
+  },
+  session: {
+    strategy: "jwt",
+  },
+  secret: process.env.NEXTAUTH_SECRET,
+};
